refactor(raffle): split store by concern — core, claim protocol, announcements
The gate's primitive-obsession verdict is per-file and the single-file
store kept tripping it however individual signatures were cleaned. The
split is also the honest shape: raffle_claim.go owns the !claim
protocol (typed outcome, Lua script, receipt decoding) and
raffle_announce.go owns the engine-side voice (auto-close, reminder
ticks, localized lane-routed posts). No behavior change.
